create uncertainty.py
